requirements.txt
phidata
python-dotenv
yfinance
groq
uvicorn
duckduckgo-search
fastapi
packaging
openai
Flask
gunicorn
requests
nltk
alpha_vantage
flask-dotenv
newsapi-python